Adding Items location update and registration location update
require more fields on join to ensure we know where members are located
to mail calgary dollars.

Now that we require these fields we can auto fill locations for ads so
added that info back in.

Also made some content changes at the top of the registration page to
clarify joining the site and steps needed to complete so that users can
get free calgary dolllars.